Generate CNF file: Minesniffer is a variant of the popular Minesweeper game. The Minesniffer game consists of an N x M board of cells (N rows and M columns). Each cell may potentially hold one mine. The task at hand is to locate all the cells with mines. Hints to locate them are available in cells that do not have mines. Specifically the cells without mines will be labeled with a numeric value between 0 and 8. The numeric value in a cell in the (i,j)th position of the board denotes the number of mines in cells adjacent to (i,j). Cells adjacent to the cell at (i,j) are those bordering it on its left, right, above, below and the corners.
Prolog Tasks:
1. The Missionaries and Cannibals Problem:
Three missionaries and three cannibals seek to cross a river. A boat is available which holds two people, and which can be navigated by any combination of      missionaries and cannibals involving one or two people. If the missionaries on either bank of the river, or “en route” in the river are outnumbered at any time by the cannibals, the cannibals will succumb to their anthropologic tendencies and do away with the missionaries. Find the simplest schedule of crossing that will permit all the missionaries and cannibals to cross the river safely
2. The Assembly Problem:
Nine parts of an electronic board have square shape, the same size and each edge of every part is marked with a letter and a plus or minus sign. The parts are to be assembled into a complete board as shown in the figure below such that the common edges have the same letter and opposite signs. Write a planner in Prolog such that the program takes ?assemble as the query and outputs how to assemble the parts, i.e. determine the locations and positions of the parts w.r.t. the current positions so that they fit together to make the complete board.
3. The Puzzling case of the water and zebra:
A situation involves five houses in a row amongst which are distributed five colors, five drinks, five nationalities, five cigarettes, and five pets subject to the following constraints:
The Englishman lives in the red house.
The Spaniard owns a dog.
The Norwegian lives in the first house
Kools are smoked in the yellow house.
Chesterfields are smoked next to where the fox is kept.
The Norwegian lives next to the blue house.
The Old Gold Smoker owns snails
The Lucky Strike smoker drinks orange juice
The Ukrainian drinks tea
The Japanese smokes Parliaments
The Kools smoker lives next to where the house is kept
Coffee is drunk in the green house
The green house is to the immediate right of the ivory house
Milk is drunk in the middle house
Mr. Watson asks Mr. Holmes: Who drinks water? Who keeps the zebra?
Elementary Mr. Watson says Mr. Holmes: -- drinks water; -- keeps zebra. What is -- ?